Move materialize functionality to duck db and rename materializedRef to materialized and materialize_config to materialize
