+++ packages/connectors/duckdb/rollup.config.mjs
@@ -14,6 +14,8 @@ export default {
     }),
   ],
   external: [
+    'path',
+    '@latitude-data/sql-compiler',
     '@latitude-data/source-manager',
     '@latitude-data/query_result',
     'duckdb-async',
