+++ packages/connectors/duckdb/src/index.ts
@@ -1,10 +1,17 @@
+import path from 'path'
 import {
   BaseConnector,
   CompiledQuery,
   ConnectorError,
   ConnectorOptions,
+  Source,
   ResolvedParam,
+  type BuildSupportedMethodsArgs,
 } from '@latitude-data/source-manager'
+import {
+  emptyMetadata,
+  type SupportedMethod,
+} from '@latitude-data/sql-compiler'
 import QueryResult, { DataType, Field } from '@latitude-data/query_result'
 import { Database, OPEN_READONLY, OPEN_READWRITE } from 'duckdb-async'
 
@@ -30,6 +37,78 @@ export default class DuckdbConnector extends BaseConnector<ConnectionParams> {
     return {
       value,
       resolvedAs: '?',
+    }
+  }
+
+  protected buildSupportedMethods(
+    buildArgs: BuildSupportedMethodsArgs,
+  ): Record<string, SupportedMethod> {
+    const supportedMethods = super.buildSupportedMethods(buildArgs)
+    const { context } = buildArgs
+
+    return {
+      ...supportedMethods,
+      materialized: {
+        requirements: {
+          interpolationPolicy: 'require',
+          interpolationMethod: 'raw',
+          requireStaticArguments: true,
+        },
+        resolve: async (referencedQuery: string) => {
+          if (typeof referencedQuery !== 'string') {
+            throw new Error('Invalid query name')
+          }
+          const fullSubQueryPath = referencedQuery.startsWith('/')
+            ? referencedQuery
+            : path.join(
+                path.dirname(context.request.queryPath),
+                referencedQuery,
+              )
+
+          if (
+            context.queriesBeingCompiled.includes(
+              fullSubQueryPath.replace(/.sql$/, ''),
+            )
+          ) {
+            throw new Error(
+              'Query reference to a parent, resulting in cyclic references.',
+            )
+          }
+
+          const refSource = (await this.source.manager.loadFromQuery(
+            fullSubQueryPath,
+          )) as Source
+          const { config, methods, sqlHash } =
+            await refSource.getMetadataFromQuery(fullSubQueryPath)
+          if (!config.materialize) {
+            throw new Error(
+              `Referenced query is not a materialized. \nYou can materialize it by adding {@config materialized_query = true} in the query content.`,
+            )
+          }
+
+          const unsupportedMethods = ['param', 'runQuery']
+          const unsupportedMethodsInQuery = Array.from(methods).filter(
+            (method) => unsupportedMethods.includes(method),
+          )
+          if (unsupportedMethodsInQuery.length > 0) {
+            const unsupportedMethodsStr = unsupportedMethodsInQuery.join(', ')
+            throw new Error(
+              `Referenced query must be static. It can not contain any of the following methods: ${unsupportedMethodsStr}`,
+            )
+          }
+
+          const storage = await this.source.manager.materializeStorage
+          const materializeUrl = await storage.getUrl({
+            sqlHash,
+            sourcePath: refSource.path,
+            queryName: referencedQuery,
+          })
+          return `read_parquet('${materializeUrl}')`
+        },
+        readMetadata: async () => {
+          return emptyMetadata()
+        },
+      },
     }
   }
