+++ packages/connectors/duckdb/src/tests/helpers.ts
