+++ packages/connectors/duckdb/tsconfig.json
@@ -2,7 +2,10 @@
   "extends": "@latitude-data/typescript/base.json",
   "compilerOptions": {
     "rootDir": "src",
-    "outDir": "dist"
+    "outDir": "dist",
+    "paths": {
+      "$/*": ["./src/*"]
+    }
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules", "dist"]
