+++ packages/connectors/duckdb/vitest.config.ts
