--- packages/connectors/materialized/.eslintrc.cjs
@@ -1,4 +1,0 @@
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-  extends: ['@latitude-data/eslint-config/library.js'],
-}
