--- packages/connectors/materialized/CHANGELOG.md
@@ -1,16 +1,0 @@
-# @latitude-data/materialized-connector
-
-## 0.0.1-next.0
-
-### Patch Changes
-
-- 26aa69d: Updated custom method definitions to new version
-- Updated dependencies [9e2dd26]
-- Updated dependencies [bd39d29]
-- Updated dependencies [aab4a4e]
-- Updated dependencies [b95b26d]
-- Updated dependencies [26aa69d]
-- Updated dependencies [5388504]
-- Updated dependencies [a8d4658]
-  - @latitude-data/source-manager@1.0.0-next.0
-  - @latitude-data/duckdb-connector@3.0.0-next.0
