--- packages/connectors/materialized/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "@latitude-data/typescript/base.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist",
-    "paths": {
-      "$/*": ["./src/*"]
-    }
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
-}
